moved the open shop button
